refactor: 로그인 관련 클래스 member에서 auth 패키지로 이동
